move test to examples
